+++ src/autocomplete/autocomplete.scss
@@ -1,83 +1,58 @@
 @import '../var-defaults.scss';
-@import '../_mixins.scss';
 @import '_attractions-theme.scss';
 
-.shown-on-focus {
+.autocomplete {
+  padding: .25em .625em .0625em;
+  box-sizing: border-box;
+  border-bottom: .0625em solid $textfield-border;
+  background-color: $textfield-bg;
+  border-top-left-radius: .375em;
+  border-top-right-radius: .375em;
+  width: 100%;
   display: flex;
-  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
 
-  & > :global(.btn) {
-    @include hide-accessible;
+  &:hover {
+    background-color: darken($textfield-bg, 3%);
   }
 
-  & > :global(.btn:focus) {
-    @include show-accessible;
-
-    margin: .5em;
-  }
-}
-
-.notice {
-  font-size: .9em;
-  margin: .5em;
-  font-weight: $thin-font-weight;
-}
-
-input + :global(.dropdown) {
-  border-radius: .5em;
-  overflow: hidden;
-}
-
-.options-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  max-height: 15em;
-  overflow-y: scroll;
-  overscroll-behavior: none;
-
-  &::-webkit-scrollbar {
-    width: .3125em;
-    height: .5em;
-    background: none;
-    margin: 1em 0;
+  &.focus {
+    background-color: darken($textfield-bg, 8%);
+    border-bottom-width: .125em;
+    border-color: lighten($main, 4%);
+    padding-bottom: 0;
   }
 
-  &::-webkit-scrollbar-thumb {
-    background: $light-contrast;
-    border-radius: .3125em;
+  :global(.autocomplete-field) {
+    flex: 1;
 
-    &:hover {
-      background: darken($light-contrast, 10%);
+    :global(input) {
+      border: none;
+      background: none;
+      outline: none;
+      min-width: 13em;
+      height: 2.8em;
+      font-family: $font;
+      width: 100%;
+      font-size: 1em;
     }
   }
 
-  :global(li:first-child) {
-    margin-top: .5em;
-  }
-
-  :global(li:last-child) {
-    margin-bottom: .5em;
-  }
-
-  .loading-options {
+  :global(.chip.no-padding) {
     display: flex;
     align-items: center;
+    padding-left: .7em;
+    margin-right: .4em;
     font-size: .95em;
-    margin: .5em 1.2em;
 
-    :global(.spinner) {
-      margin-right: .7em;
-      margin-bottom: .3em;
+    :global(.btn) {
+      margin-left: .3em;
     }
-  }
-
-  .more-options {
-    margin-left: .5em;
-    margin-right: .5em;
 
     :global(svg) {
-      margin-right: .5em;
+      width: 1.2em;
+      height: 1.2em;
     }
   }
 }
